Fix: stabilize story e.2 inbound SMS ATDD suite
