+++ _bmad-output/test-artifacts/test-review-validation-report.md
@@ -1,0 +1,40 @@
+# Test Review Validation Report (Story e.2)
+
+- Workflow: `testarch-test-review`
+- Date: 2026-03-03
+- Scope: `single`
+- Target: `tests/api/platform/e-2-inbound-sms-processing-with-active-thread-ensure.atdd.api.spec.ts` (split modules reviewed)
+
+## Section Results
+
+- Prerequisites: PASS
+- Test file discovery and readability: PASS
+- Knowledge base loading (required fragments): PASS
+- Context gathering (story + framework): PASS
+- Step 1 (context loading): PASS
+- Step 2 (test parsing and metadata extraction): PASS
+- Step 3 (5-dimension quality evaluation): PASS
+- Step 3F (weighted aggregation): PASS
+- Step 4 (report generation): PASS
+- Checklist validation pass: PASS
+
+## Checklist Summary
+
+- Quality criteria evaluated: PASS
+- Score and grade computed from weighted dimensions: PASS
+- Severity classification applied: PASS
+- Actionable recommendations with locations: PASS
+- Acceptance criteria mapping included: PASS
+- Knowledge base references included: PASS
+- Temp review artifacts persisted under `_bmad-output/test-artifacts/test-review-temp`: PASS
+- CLI session cleanup requirement: PASS (no CLI browser session opened for API-only scope)
+
+## Notes
+
+- Rerun reflects post-fix state: replay-suppression now includes persisted timeline cardinality checks and story e.2 tests are split into focused modules.
+- Optional outputs (`inline comments`, `quality badge`, `story append`) were not enabled by this run.
+- No blocking gaps detected.
+
+## Validation Verdict
+
+PASS with no blocking gaps.
