+++ tests/api/platform/e-2-inbound-sms-processing-with-active-thread-ensure.atdd.api.spec.ts
@@ -1,595 +1,5 @@
-import { apiRequest } from '../../support/helpers/apiClient';
-import { test, expect } from '../../support/fixtures/connectShyftStoryE2.fixture';
-import {
-  deterministicProviderEventId,
-  deterministicToken,
-} from '../../support/utils/deterministicTestIds';
-import {
-  buildSignedWebhookHeaders,
-  buildSmsWebhookPayload,
-  hasRequiredEnvelopeKeys,
-} from '../../support/helpers/connectShyftWebhookTestHelpers';
-
-const UUID_PATTERN = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
-
-const expectsDurableSideEffects = (tenantId: string, threadId: string): boolean => {
-  return UUID_PATTERN.test(tenantId) && UUID_PATTERN.test(threadId);
-};
-
-test.describe(
-  'Story e.2 Inbound SMS Processing with Active-Thread Ensure (ATDD API RED)',
-  () => {
-    test(
-      '[E2-ATDD-API-001][P0] mapped inbound sms ensures one active thread for tenant-orgUnit-neighbor and appends inbound message artifact @P0',
-      async ({
-        request,
-        storyE2Context,
-        storyE2AdminHeaders,
-        storyE2NumberMappingPayload,
-      }, testInfo) => {
-        const mappingResponse = await apiRequest(request, {
-          method: 'POST',
-          path: storyE2Context.paths.numbersCollection,
-          headers: storyE2AdminHeaders,
-          data: storyE2NumberMappingPayload,
-        });
-        expect([200, 201]).toContain(mappingResponse.status());
-
-        const webhookPayload = {
-          ...buildSmsWebhookPayload({
-            providerKey: storyE2Context.providers.enabledPrimary,
-            to: storyE2Context.numbers.mappedInbound,
-            from: storyE2Context.numbers.mappedOutbound,
-            providerMessageId: `msg-e2-atdd-api-001-${deterministicToken(testInfo, 'e2-atdd-api-001-message')}`,
-            providerEventId: deterministicProviderEventId(
-              'provider-event-e2-atdd-api',
-              testInfo,
-              'mapped-ensure-append',
-            ),
-          }),
-          neighborId: `${storyE2Context.neighborIds.inboundCreate}-new-${deterministicToken(testInfo, 'e2-atdd-api-001-neighbor')}`,
-        };
-
-        const webhookResponse = await apiRequest(request, {
-          method: 'POST',
-          path: storyE2Context.paths.inboundWebhook,
-          headers: {
-            ...storyE2AdminHeaders,
-            ...buildSignedWebhookHeaders(webhookPayload, testInfo, 'e2-atdd-api-001'),
-          },
-          data: webhookPayload,
-        });
-
-        expect(webhookResponse.status()).toBe(200);
-        const webhookBody = await webhookResponse.json();
-        expect(hasRequiredEnvelopeKeys(webhookBody)).toBe(true);
-        expect(webhookBody).toMatchObject({
-          ok: true,
-          code: 'CONNECTSHYFT_WEBHOOK_ACCEPTED',
-          data: {
-            correlation: {
-              source: 'number_mapping',
-              deterministic: true,
-              tenantId: storyE2Context.tenantId,
-              orgUnitId: storyE2Context.orgUnitId,
-              neighborId: webhookPayload.neighborId,
-            },
-            replaySafe: {
-              duplicate: false,
-              suppressedDomainWrites: false,
-              dedupeKey: expect.stringContaining('provider-event:'),
-            },
-            thread: {
-              tenantId: storyE2Context.tenantId,
-              orgUnitId: storyE2Context.orgUnitId,
-              neighborId: webhookPayload.neighborId,
-              state: 'UNCLAIMED',
-            },
-            lifecycle: {
-              ensuredActiveThread: true,
-            },
-            inboundMessageArtifact: {
-              channel: 'sms',
-              direction: 'inbound',
-              providerEventId: expect.any(String),
-              body: expect.any(String),
-            },
-            timeline: {
-              eventName: storyE2Context.eventNames.inboundSmsAppended,
-              routingDecision: 'accepted',
-            },
-            sideEffects: {
-              canonicalEventPersisted: true,
-            },
-          },
-        });
-
-        const resolvedThreadId = String(webhookBody?.data?.thread?.threadId || webhookBody?.data?.threadId || '');
-        expect(resolvedThreadId.length).toBeGreaterThan(0);
-        const durableSideEffects = expectsDurableSideEffects(storyE2Context.tenantId, resolvedThreadId);
-        expect(webhookBody?.data?.sideEffects).toMatchObject({
-          canonicalEventPersisted: true,
-          auditPersisted: durableSideEffects,
-          outboxPersisted: durableSideEffects,
-        });
-        expect(webhookBody?.data?.transaction).toMatchObject({
-          atomic: durableSideEffects,
-          auditPersisted: durableSideEffects,
-          outboxPersisted: durableSideEffects,
-        });
-        if (durableSideEffects) {
-          expect(webhookBody?.data?.audit?.eventName).toBe(storyE2Context.eventNames.inboundSmsAppended);
-          expect(webhookBody?.data?.outbox?.eventName).toBe(storyE2Context.eventNames.inboundSmsAppended);
-        } else {
-          expect(webhookBody.data).not.toHaveProperty('audit');
-          expect(webhookBody.data).not.toHaveProperty('outbox');
-        }
-      },
-    );
-
-    test(
-      '[E2-ATDD-API-002][P0] existing active thread is reused without duplicate creation and ordering remains deterministic @P0',
-      async ({
-        request,
-        storyE2Context,
-        storyE2OperatorHeaders,
-        storyE2AdminHeaders,
-        storyE2NumberMappingPayload,
-        storyE2EnsurePayload,
-      }, testInfo) => {
-        await apiRequest(request, {
-          method: 'POST',
-          path: storyE2Context.paths.numbersCollection,
-          headers: storyE2AdminHeaders,
-          data: storyE2NumberMappingPayload,
-        });
-
-        const ensureResponse = await apiRequest(request, {
-          method: 'POST',
-          path: storyE2Context.paths.threads,
-          headers: storyE2OperatorHeaders,
-          data: storyE2EnsurePayload,
-        });
-        expect(ensureResponse.status()).toBe(201);
-        const ensureBody = await ensureResponse.json();
-        const existingThreadId = String(ensureBody?.data?.thread?.threadId || '');
-        expect(existingThreadId.length).toBeGreaterThan(0);
-
-        const webhookPayload = {
-          ...buildSmsWebhookPayload({
-            providerKey: storyE2Context.providers.enabledPrimary,
-            to: storyE2Context.numbers.mappedInbound,
-            from: storyE2Context.numbers.mappedOutbound,
-            providerMessageId: `msg-e2-atdd-api-002-${deterministicToken(testInfo, 'e2-atdd-api-002-message')}`,
-            providerEventId: deterministicProviderEventId(
-              'provider-event-e2-atdd-api',
-              testInfo,
-              'existing-thread-reuse',
-            ),
-          }),
-          neighborId: storyE2Context.neighborIds.existingActive,
-        };
-
-        const webhookResponse = await apiRequest(request, {
-          method: 'POST',
-          path: storyE2Context.paths.inboundWebhook,
-          headers: {
-            ...storyE2AdminHeaders,
-            ...buildSignedWebhookHeaders(webhookPayload, testInfo, 'e2-atdd-api-002'),
-          },
-          data: webhookPayload,
-        });
-        expect(webhookResponse.status()).toBe(200);
-        const webhookBody = await webhookResponse.json();
-        expect(hasRequiredEnvelopeKeys(webhookBody)).toBe(true);
-        expect(webhookBody).toMatchObject({
-          ok: true,
-          code: 'CONNECTSHYFT_WEBHOOK_ACCEPTED',
-          data: {
-            thread: {
-              threadId: existingThreadId,
-              state: 'UNCLAIMED',
-            },
-            lifecycle: {
-              ensuredActiveThread: true,
-              createdNewThread: false,
-              reusedThreadId: existingThreadId,
-            },
-            timeline: {
-              deterministicOrdering: true,
-            },
-          },
-        });
-
-        const detailResponse = await apiRequest(request, {
-          method: 'GET',
-          path: `${storyE2Context.paths.threads}/${existingThreadId}`,
-          headers: storyE2OperatorHeaders,
-        });
-        expect(detailResponse.status()).toBe(200);
-        const detailBody = await detailResponse.json();
-        const timeline = detailBody?.data?.thread?.timeline as Array<{
-          eventName: string;
-          occurredAtUtc: string;
-          eventId: string;
-        }>;
-        const inboundSmsEvents = timeline.filter(
-          (entry) => entry.eventName === storyE2Context.eventNames.inboundSmsAppended,
-        );
-        expect(inboundSmsEvents.length).toBeGreaterThan(0);
-        const sorted = [...inboundSmsEvents].sort((left, right) => {
-          const timeDelta =
-            new Date(left.occurredAtUtc).getTime() -
-            new Date(right.occurredAtUtc).getTime();
-          return timeDelta !== 0 ? timeDelta : left.eventId.localeCompare(right.eventId);
-        });
-        expect(inboundSmsEvents).toEqual(sorted);
-      },
-    );
-
-    test(
-      '[E2-ATDD-API-003][P0] duplicate inbound webhook deliveries suppress duplicate timeline entries via provider event replay identity @P0',
-      async ({
-        request,
-        storyE2Context,
-        storyE2AdminHeaders,
-        storyE2NumberMappingPayload,
-      }, testInfo) => {
-        await apiRequest(request, {
-          method: 'POST',
-          path: storyE2Context.paths.numbersCollection,
-          headers: storyE2AdminHeaders,
-          data: storyE2NumberMappingPayload,
-        });
-
-        const providerEventId = deterministicProviderEventId(
-          'provider-event-e2-atdd-api',
-          testInfo,
-          'duplicate-suppression',
-        );
-        const webhookPayload = {
-          ...buildSmsWebhookPayload({
-            providerKey: storyE2Context.providers.enabledPrimary,
-            to: storyE2Context.numbers.mappedInbound,
-            from: storyE2Context.numbers.mappedOutbound,
-            providerMessageId: `msg-e2-atdd-api-003-${deterministicToken(testInfo, 'e2-atdd-api-003-message')}`,
-            providerEventId,
-          }),
-          neighborId: storyE2Context.neighborIds.inboundDuplicate,
-        };
-
-        const firstResponse = await apiRequest(request, {
-          method: 'POST',
-          path: storyE2Context.paths.inboundWebhook,
-          headers: {
-            ...storyE2AdminHeaders,
-            ...buildSignedWebhookHeaders(webhookPayload, testInfo, 'e2-atdd-api-003-first'),
-          },
-          data: webhookPayload,
-        });
-        const duplicateResponse = await apiRequest(request, {
-          method: 'POST',
-          path: storyE2Context.paths.inboundWebhook,
-          headers: {
-            ...storyE2AdminHeaders,
-            ...buildSignedWebhookHeaders(webhookPayload, testInfo, 'e2-atdd-api-003-duplicate'),
-          },
-          data: webhookPayload,
-        });
-
-        expect(firstResponse.status()).toBe(200);
-        expect(duplicateResponse.status()).toBe(200);
-
-        const firstBody = await firstResponse.json();
-        const duplicateBody = await duplicateResponse.json();
-        expect(hasRequiredEnvelopeKeys(firstBody)).toBe(true);
-        expect(hasRequiredEnvelopeKeys(duplicateBody)).toBe(true);
-
-        expect(firstBody).toMatchObject({
-          ok: true,
-          code: 'CONNECTSHYFT_WEBHOOK_ACCEPTED',
-          data: {
-            replaySafe: {
-              duplicate: false,
-              suppressedDomainWrites: false,
-              dedupeKey: `provider-event:${providerEventId.toLowerCase()}`,
-            },
-          },
-        });
-        expect(duplicateBody).toMatchObject({
-          ok: true,
-          code: 'CONNECTSHYFT_WEBHOOK_ACCEPTED',
-          data: {
-            replaySafe: {
-              duplicate: true,
-              suppressedDomainWrites: true,
-              dedupeKey: `provider-event:${providerEventId.toLowerCase()}`,
-            },
-            sideEffects: {
-              lifecycleMutationApplied: false,
-              canonicalEventPersisted: false,
-              outboxPersisted: false,
-            },
-          },
-        });
-        expect(duplicateBody.data).not.toHaveProperty('timeline');
-        expect(duplicateBody.data).not.toHaveProperty('audit');
-        expect(duplicateBody.data).not.toHaveProperty('outbox');
-      },
-    );
-
-    test(
-      '[E2-ATDD-API-004][P1] create-and-append path writes inbound message artifact atomically with audit and outbox metadata @P1',
-      async ({
-        request,
-        storyE2Context,
-        storyE2OperatorHeaders,
-        storyE2AdminHeaders,
-        storyE2NumberMappingPayload,
-      }, testInfo) => {
-        await apiRequest(request, {
-          method: 'POST',
-          path: storyE2Context.paths.numbersCollection,
-          headers: storyE2AdminHeaders,
-          data: storyE2NumberMappingPayload,
-        });
-
-        const webhookPayload = {
-          ...buildSmsWebhookPayload({
-            providerKey: storyE2Context.providers.enabledPrimary,
-            to: storyE2Context.numbers.mappedInbound,
-            from: storyE2Context.numbers.mappedOutbound,
-            providerMessageId: `msg-e2-atdd-api-004-${deterministicToken(testInfo, 'e2-atdd-api-004-message')}`,
-            providerEventId: deterministicProviderEventId(
-              'provider-event-e2-atdd-api',
-              testInfo,
-              'atomic-create-append',
-            ),
-          }),
-          neighborId: `${storyE2Context.neighborIds.inboundCreate}-atomic-${deterministicToken(testInfo, 'e2-atdd-api-004-neighbor')}`,
-        };
-
-        const response = await apiRequest(request, {
-          method: 'POST',
-          path: storyE2Context.paths.inboundWebhook,
-          headers: {
-            ...storyE2AdminHeaders,
-            ...buildSignedWebhookHeaders(webhookPayload, testInfo, 'e2-atdd-api-004'),
-          },
-          data: webhookPayload,
-        });
-        expect(response.status()).toBe(200);
-        const body = await response.json();
-        expect(hasRequiredEnvelopeKeys(body)).toBe(true);
-        expect(body).toMatchObject({
-          ok: true,
-          code: 'CONNECTSHYFT_WEBHOOK_ACCEPTED',
-          data: {
-            lifecycle: {
-              ensuredActiveThread: true,
-            },
-            inboundMessageArtifact: {
-              artifactId: expect.any(String),
-              channel: 'sms',
-              direction: 'inbound',
-            },
-            sideEffects: {
-              canonicalEventPersisted: true,
-            },
-          },
-        });
-
-        const resolvedThreadId = String(body?.data?.thread?.threadId || body?.data?.threadId || '');
-        expect(resolvedThreadId.length).toBeGreaterThan(0);
-        const durableSideEffects = expectsDurableSideEffects(storyE2Context.tenantId, resolvedThreadId);
-        expect(body?.data?.sideEffects).toMatchObject({
-          canonicalEventPersisted: true,
-          auditPersisted: durableSideEffects,
-          outboxPersisted: durableSideEffects,
-        });
-        expect(body?.data?.transaction).toMatchObject({
-          atomic: durableSideEffects,
-          auditPersisted: durableSideEffects,
-          outboxPersisted: durableSideEffects,
-        });
-        if (durableSideEffects) {
-          expect(body?.data?.audit?.eventName).toBe(storyE2Context.eventNames.inboundSmsAppended);
-          expect(body?.data?.outbox?.eventName).toBe(storyE2Context.eventNames.inboundSmsAppended);
-        } else {
-          expect(body.data).not.toHaveProperty('audit');
-          expect(body.data).not.toHaveProperty('outbox');
-        }
-
-        const detailResponse = await apiRequest(request, {
-          method: 'GET',
-          path: `${storyE2Context.paths.threads}/${resolvedThreadId}`,
-          headers: storyE2OperatorHeaders,
-        });
-        expect(detailResponse.status()).toBe(200);
-      },
-    );
-
-    test(
-      '[E2-ATDD-API-006][P1] inbound sms without resolvable neighbor context is refused and does not mutate timeline @P1',
-      async ({
-        request,
-        storyE2Context,
-        storyE2AdminHeaders,
-        storyE2NumberMappingPayload,
-      }, testInfo) => {
-        const mappingResponse = await apiRequest(request, {
-          method: 'POST',
-          path: storyE2Context.paths.numbersCollection,
-          headers: storyE2AdminHeaders,
-          data: storyE2NumberMappingPayload,
-        });
-        expect([200, 201]).toContain(mappingResponse.status());
-
-        const webhookPayload = buildSmsWebhookPayload({
-          providerKey: storyE2Context.providers.enabledPrimary,
-          to: storyE2Context.numbers.mappedInbound,
-          from: storyE2Context.numbers.mappedOutbound,
-          providerMessageId: `msg-e2-atdd-api-006-${deterministicToken(testInfo, 'e2-atdd-api-006-message')}`,
-          providerEventId: deterministicProviderEventId(
-            'provider-event-e2-atdd-api',
-            testInfo,
-            'neighbor-unresolved',
-          ),
-        });
-
-        const webhookResponse = await apiRequest(request, {
-          method: 'POST',
-          path: storyE2Context.paths.inboundWebhook,
-          headers: {
-            ...storyE2AdminHeaders,
-            ...buildSignedWebhookHeaders(webhookPayload, testInfo, 'e2-atdd-api-006'),
-          },
-          data: webhookPayload,
-        });
-
-        expect(webhookResponse.status()).toBe(200);
-        const webhookBody = await webhookResponse.json();
-        expect(hasRequiredEnvelopeKeys(webhookBody)).toBe(true);
-        expect(webhookBody).toMatchObject({
-          ok: false,
-          code: 'CONNECTSHYFT_WEBHOOK_NEIGHBOR_UNRESOLVED',
-          data: {
-            sideEffects: {
-              lifecycleMutationApplied: false,
-              canonicalEventPersisted: false,
-              outboxPersisted: false,
-            },
-            timelineOutcome: {
-              eventName: null,
-              routingDecision: 'refused',
-            },
-            reason: 'neighbor_unresolved',
-          },
-        });
-      },
-    );
-
-    test(
-      '[E2-ATDD-API-005][P0] concurrent inbound deliveries converge to a single active thread identity while appending both artifacts @P0',
-      async ({
-        request,
-        storyE2Context,
-        storyE2OperatorHeaders,
-        storyE2AdminHeaders,
-        storyE2NumberMappingPayload,
-      }, testInfo) => {
-        await apiRequest(request, {
-          method: 'POST',
-          path: storyE2Context.paths.numbersCollection,
-          headers: storyE2AdminHeaders,
-          data: storyE2NumberMappingPayload,
-        });
-
-        const concurrentNeighborId = `${storyE2Context.neighborIds.inboundCreate}-concurrent-${deterministicToken(testInfo, 'e2-atdd-api-005-neighbor')}`;
-        const webhookPayloadA = {
-          ...buildSmsWebhookPayload({
-            providerKey: storyE2Context.providers.enabledPrimary,
-            to: storyE2Context.numbers.mappedInbound,
-            from: storyE2Context.numbers.mappedOutbound,
-            providerMessageId: `msg-e2-atdd-api-005-a-${deterministicToken(testInfo, 'e2-atdd-api-005-message-a')}`,
-            providerEventId: deterministicProviderEventId(
-              'provider-event-e2-atdd-api',
-              testInfo,
-              'concurrent-a',
-            ),
-          }),
-          neighborId: concurrentNeighborId,
-        };
-        const webhookPayloadB = {
-          ...buildSmsWebhookPayload({
-            providerKey: storyE2Context.providers.enabledPrimary,
-            to: storyE2Context.numbers.mappedInbound,
-            from: storyE2Context.numbers.mappedOutbound,
-            providerMessageId: `msg-e2-atdd-api-005-b-${deterministicToken(testInfo, 'e2-atdd-api-005-message-b')}`,
-            providerEventId: deterministicProviderEventId(
-              'provider-event-e2-atdd-api',
-              testInfo,
-              'concurrent-b',
-            ),
-          }),
-          neighborId: concurrentNeighborId,
-        };
-
-        const [firstResponse, secondResponse] = await Promise.all([
-          apiRequest(request, {
-            method: 'POST',
-            path: storyE2Context.paths.inboundWebhook,
-            headers: {
-              ...storyE2AdminHeaders,
-              ...buildSignedWebhookHeaders(webhookPayloadA, testInfo, 'e2-atdd-api-005-a'),
-            },
-            data: webhookPayloadA,
-          }),
-          apiRequest(request, {
-            method: 'POST',
-            path: storyE2Context.paths.inboundWebhook,
-            headers: {
-              ...storyE2AdminHeaders,
-              ...buildSignedWebhookHeaders(webhookPayloadB, testInfo, 'e2-atdd-api-005-b'),
-            },
-            data: webhookPayloadB,
-          }),
-        ]);
-
-        expect(firstResponse.status()).toBe(200);
-        expect(secondResponse.status()).toBe(200);
-
-        const firstBody = await firstResponse.json();
-        const secondBody = await secondResponse.json();
-        expect(hasRequiredEnvelopeKeys(firstBody)).toBe(true);
-        expect(hasRequiredEnvelopeKeys(secondBody)).toBe(true);
-        expect(firstBody).toMatchObject({
-          ok: true,
-          code: 'CONNECTSHYFT_WEBHOOK_ACCEPTED',
-          data: {
-            lifecycle: {
-              ensuredActiveThread: true,
-            },
-            replaySafe: {
-              duplicate: false,
-            },
-          },
-        });
-        expect(secondBody).toMatchObject({
-          ok: true,
-          code: 'CONNECTSHYFT_WEBHOOK_ACCEPTED',
-          data: {
-            lifecycle: {
-              ensuredActiveThread: true,
-            },
-            replaySafe: {
-              duplicate: false,
-            },
-          },
-        });
-
-        const firstThreadId = String(firstBody?.data?.thread?.threadId || firstBody?.data?.threadId || '');
-        const secondThreadId = String(secondBody?.data?.thread?.threadId || secondBody?.data?.threadId || '');
-        expect(firstThreadId.length).toBeGreaterThan(0);
-        expect(secondThreadId.length).toBeGreaterThan(0);
-        expect(firstThreadId).toBe(secondThreadId);
-
-        const detailResponse = await apiRequest(request, {
-          method: 'GET',
-          path: `${storyE2Context.paths.threads}/${firstThreadId}`,
-          headers: storyE2OperatorHeaders,
-        });
-        expect(detailResponse.status()).toBe(200);
-
-        const detailBody = await detailResponse.json();
-        const timeline = detailBody?.data?.thread?.timeline as Array<{
-          eventName: string;
-        }>;
-        const inboundSmsEvents = timeline.filter(
-          (entry) => entry.eventName === storyE2Context.eventNames.inboundSmsAppended,
-        );
-        expect(inboundSmsEvents.length).toBeGreaterThanOrEqual(2);
-      },
-    );
-  },
-);
+// Story e.2 ATDD API entrypoint kept for command compatibility.
+// Tests are split into focused case modules for maintainability.
+import './e-2-inbound-sms-processing-with-active-thread-ensure.atdd.api.core.cases';
+import './e-2-inbound-sms-processing-with-active-thread-ensure.atdd.api.replay-refusal.cases';
+import './e-2-inbound-sms-processing-with-active-thread-ensure.atdd.api.concurrency.cases';
